Get strategy analysis utility working, provide capability to view valid, invalid, and malformed strats
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/branches/strategy-loading@23964 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/query/param/AnswerParam.java
@@ -227,7 +227,7 @@ public class AnswerParam extends Param {
 
   public static long toStepId(String stableValue) {
     if (FormatUtil.isInteger(stableValue)) {
-      return Integer.parseInt(stableValue);
+      return Long.parseLong(stableValue);
     }
     throw new WdkRuntimeException("AnswerParam value '" + stableValue + "' is not an integer.");
   }
